requirements.txt
aiohttp==3.8.4
aiofiles==0.8.0
aio_pika==8.2.3
pyyaml==6.0
